Rename dev to cli
