+++ .github/workflows/ci-testing.yml
@@ -29,8 +29,8 @@ jobs:
     - name: Set oldest dependencies
       if: matrix.requires == 'oldest'
       run: |
-        pip install -e '.[dev]'
-        python -m lightning_utilities.dev requirements set-oldest
+        pip install -e '.[cli]'
+        python -m lightning_utilities.cli requirements set-oldest
 
     - uses: ./.github/actions/cache
       with:
