+++ src/lightning_utilities/cli/__main__.py
@@ -4,7 +4,7 @@
 #
 import fire
 
-from lightning_utilities.dev.dependencies import prune_pkgs_in_requirements, replace_oldest_ver
+from lightning_utilities.cli.dependencies import prune_pkgs_in_requirements, replace_oldest_ver
 
 
 def main() -> None:
